refactor: split infoforms UI builders into infoforms_ui.go
